+++ preview/MsixCore/MsixCoreInstaller/FileTypeAssociation.cpp
@@ -13,10 +13,12 @@
 #include <filesystem>
 #include <fstream>
 
-// GdiPlus.h requires a definiton for min and max. Use std namespace *BEFORE* including it.
-using namespace std;
+// MSIXWindows.hpp defines NOMINMAX and undefines min and max because we want to use std::min/std::max from <algorithm>
+// GdiPlus.h requires a definiton for min and max. We can't use namespace std because c++17 defines std::byte, which conflicts with ::byte
+#define max std::max
+#define min std::min
 #include <GdiPlus.h>
-
+using namespace std;
 using namespace MsixCoreLib;
 
 const PCWSTR FileTypeAssociation::HandlerName = L"FileTypeAssociation";
